refactor(gateway): extract message-convert and usage from llm-forward/util
