+++ packages/xinity-ai-gateway/src/llm-forward/util.ts
@@ -1,234 +1,9 @@
-import { logChatSync, logChatStream, type ChatSyncData, type ChatStreamData } from "../callLogger";
-import { recordTokenUsage, recordModelRequest, recordBackendError } from "../metrics";
-import { recordUsageEvent } from "../usageRecorder";
-import type { AuthResult } from "./auth";
-import type { ModelMessage, ImagePart, TextPart } from "ai";
-import type { ApiCallInputMessage } from "common-db";
+import { recordBackendError } from "../metrics";
 import { BLOCKED_REQUEST_PARAM_PREFIXES } from "xinity-infoserver";
-import { rootLogger } from "../logger";
 
-const log = rootLogger.child({ name: "util" });
-
-// ---------------------------------------------------------------------------
-// Message conversion (OpenAI wire format → AI SDK ModelMessage)
-// ---------------------------------------------------------------------------
-
-/**
- * Convert messages from OpenAI wire format to AI SDK ModelMessage format.
- *
- * Handles:
- * - `image_url` content parts → AI SDK `{ type: "image", image }` parts
- * - Assistant messages with `tool_calls` → AI SDK tool-call content parts
- * - `tool` role messages → AI SDK tool-result content parts
- */
-export function toModelMessages(messages: ApiCallInputMessage[]): ModelMessage[] {
-  // Build a toolCallId → toolName map from assistant messages so that
-  // tool-result messages (which lack toolName in OpenAI format) can be mapped.
-  const toolCallNameMap = new Map<string, string>();
-  for (const msg of messages) {
-    const raw = msg as Record<string, unknown>;
-    if (raw.role === "assistant" && Array.isArray(raw.tool_calls)) {
-      for (const tc of raw.tool_calls as Array<{ id?: string; function?: { name?: string } }>) {
-        if (tc.id && tc.function?.name) toolCallNameMap.set(tc.id, tc.function.name);
-      }
-    }
-  }
-
-  return messages.map((msg) => {
-    const raw = msg as Record<string, unknown>;
-
-    // Assistant messages with tool_calls
-    if (raw.role === "assistant" && Array.isArray(raw.tool_calls)) {
-      const parts: Array<Record<string, unknown>> = [];
-      // Preserve any text content alongside tool calls
-      if (typeof raw.content === "string" && raw.content) {
-        parts.push({ type: "text", text: raw.content });
-      }
-      for (const tc of raw.tool_calls as Array<{ id: string; type: string; function: { name: string; arguments: string } }>) {
-        if (tc.type !== "function") continue;
-        let args: unknown;
-        try { args = JSON.parse(tc.function.arguments); } catch { args = {}; }
-        parts.push({
-          type: "tool-call",
-          toolCallId: tc.id,
-          toolName: tc.function.name,
-          input: args,
-        });
-      }
-      return { role: "assistant", content: parts } as unknown as ModelMessage;
-    }
-
-    // Tool result messages
-    if (raw.role === "tool" && typeof raw.tool_call_id === "string") {
-      const resultValue = typeof raw.content === "string" ? raw.content : JSON.stringify(raw.content);
-      return {
-        role: "tool",
-        content: [{
-          type: "tool-result",
-          toolCallId: raw.tool_call_id,
-          toolName: toolCallNameMap.get(raw.tool_call_id as string) ?? "",
-          output: { type: "text", value: resultValue },
-        }],
-      } as unknown as ModelMessage;
-    }
-
-    // Standard messages: convert image_url parts
-    if (typeof msg.content === "string" || !Array.isArray(msg.content)) {
-      return msg as ModelMessage;
-    }
-    const content = (msg.content as Array<{ type: string; text?: string; image_url?: { url: string } }>).flatMap<TextPart | ImagePart>((part) => {
-      if (part.type !== "image_url" || !part.image_url) {
-        return [part as TextPart];
-      }
-      // The AI SDK accepts data URIs and URL strings directly as image.image
-      return [{ type: "image", image: part.image_url.url }];
-    });
-    return { ...msg, content } as ModelMessage;
-  });
-}
-
-/** Usage data accepted from both AI SDK (inputTokens) and OpenAI wire format (prompt_tokens). */
-export type UsageData = {
-  inputTokens?: number;
-  outputTokens?: number;
-  prompt_tokens?: number;
-  completion_tokens?: number;
-  total_tokens?: number;
-};
-
-function normalizeInputTokens(usage: UsageData): number {
-  return usage.inputTokens ?? usage.prompt_tokens ?? 0;
-}
-
-function normalizeOutputTokens(usage: UsageData): number {
-  return usage.outputTokens ?? usage.completion_tokens ?? 0;
-}
-
-export type RecordUsageContext = {
-  usage: UsageData | null | undefined;
-  auth: AuthResult;
-  modelInfo: { model: string; nodeId?: string | null };
-  callStartTime: number;
-  /**
-   * Per-request override for whether the call body is stored:
-   *   `undefined` defers to the API key's `collectData` policy,
-   *   `true`/`false` ignore the policy and force the outcome.
-   */
-  logCalls?: boolean;
-  /** The deployment's public specifier, used for per-deployment metrics. */
-  deployment?: string;
-};
-
-/** Record token metrics and a usage event. Shared by all endpoint types. */
-export const recordUsage = ({
-  usage,
-  auth,
-  modelInfo,
-  callStartTime,
-  logCalls,
-  deployment,
-}: RecordUsageContext): boolean => {
-  const durationMs = Date.now() - callStartTime;
-  recordTokenUsage(modelInfo.model, auth.keyId, usage, { deployment, durationMs });
-  recordModelRequest(modelInfo.model, true);
-  if (!usage) {
-    return false;
-  }
-
-  const shouldLog = logCalls ?? auth.collectData;
-
-  recordUsageEvent({
-    organizationId: auth.orgId,
-    applicationId: auth.applicationId,
-    apiKeyId: auth.keyId,
-    model: modelInfo.model,
-    inputTokens: normalizeInputTokens(usage),
-    outputTokens: normalizeOutputTokens(usage),
-    duration: durationMs,
-    logged: shouldLog,
-    nodeId: modelInfo.nodeId ?? null,
-    success: true,
-  });
-
-  return shouldLog;
-};
-
-export type FailedRequestContext = {
-  auth: AuthResult;
-  modelInfo: { model: string; nodeId?: string | null };
-  callStartTime: number;
-};
-
-/** Record a usage event for a request that failed after a node was selected. */
-export function recordFailedRequest({ auth, modelInfo, callStartTime }: FailedRequestContext): void {
-  recordModelRequest(modelInfo.model, false);
-  recordUsageEvent({
-    organizationId: auth.orgId,
-    applicationId: auth.applicationId,
-    apiKeyId: auth.keyId,
-    model: modelInfo.model,
-    inputTokens: 0,
-    outputTokens: 0,
-    duration: Date.now() - callStartTime,
-    logged: false,
-    nodeId: modelInfo.nodeId ?? null,
-    success: false,
-  });
-}
-
-type UsageLogContextBase = {
-  usage: UsageData | null | undefined;
-  auth: AuthResult;
-  modelInfo: { model: string };
-  modelSpecifier: string;
-  inputMessages: ApiCallInputMessage[];
-  callStartTime: number;
-  logCalls?: boolean;
-  metadata?: Record<string, unknown>;
-};
-
-export type UsageLogContext = UsageLogContextBase & (
-  | { stream: true; outputData: ChatStreamData }
-  | { stream: false; outputData: ChatSyncData }
-);
-
-export const logChatUsage = ({
-  usage,
-  outputData,
-  stream,
-  auth,
-  modelInfo,
-  modelSpecifier,
-  inputMessages,
-  callStartTime,
-  logCalls,
-  metadata,
-}: UsageLogContext) => {
-  const shouldLog = recordUsage({ usage, auth, modelInfo, callStartTime, logCalls, deployment: modelSpecifier });
-  if (!shouldLog) {
-    return;
-  }
-
-  const commonFields = {
-    keyId: auth.keyId,
-    applicationId: auth.applicationId,
-    organizationId: auth.orgId,
-    modelSpecifier,
-    durationInMS: Date.now() - callStartTime,
-    inputMessages,
-    metadata,
-  };
-
-  if (stream) {
-    logChatStream({ ...commonFields, data: outputData }).catch((err) => {
-      log.error({ err }, "logChatStream error");
-    });
-  } else {
-    logChatSync({ ...commonFields, data: outputData }).catch((err) => {
-      log.error({ err }, "logChatSync error");
-    });
-  }
-};
+export { toModelMessages } from "./message-convert";
+export { recordUsage, recordFailedRequest, logChatUsage } from "./usage";
+export type { UsageData, RecordUsageContext, FailedRequestContext, UsageLogContext } from "./usage";
 
 // ---------------------------------------------------------------------------
 // SSE streaming helpers
